logs: renamed and moved to util pkg the method to get go routines dump.
